Move SHAUtils to util/sha_utils.py
